Clippy: casting to the same type is unnecessary (f32 -> f32)

+++ src/stats.rs
@@ -74,7 +74,7 @@ pub fn sample_entropy(m: usize, r: f32, data: &Vec<f32>) -> f32 {
 
 /// Vectorized one liner for computing the mean of a vector.
 pub fn mean(data: &Vec<f32>) -> f32 {
-    data.iter().sum::<f32>() as f32 / data.len() as f32
+    data.iter().sum::<f32>() / data.len() as f32
 }
 
 /// Vectorized read-only code that computes standard deviation.
